+++ pl.put.poznan/src/reward/Reward.java
@@ -1,5 +1,5 @@
 package reward;
 
 public interface Reward {
-    int validate() throws RewardIsNotValidException;
+    double validate() throws RewardIsNotValidException;
 }
